requirements.txt
fastapi~=0.115.12
uvicorn
jinja2
pillow
python-multipart